seeker: backend-owned recording policy + reconcile loop (the missed-ISS-pass fix)
The main cam silently didn't record a perfect ISS pass. Two bugs chained:
a bit-depth relaunch ended the cam's stream, so its recorder finalized and
exited by design -- and the GUI's edge-triggered rec_sent guard then latched
shut (reality never matched its memo), silently swallowing every later
'record on' for that role: auto-record, Record now, reconnect, everything.

Recording POLICY now lives in the backend: record_manual[role] +
record_auto[role] (record-while-tracking; turns on when a role switches to
a real camera). Every control-loop tick computes desire = manual or (auto
and tracking) and reconciles: want recording and no recorder -> start one;
don't want it but have one -> close its pipe and forget it. One brake:
never spawn within 2 s of the last spawn for that camera (crash-loop
protection), cleared whenever recording isn't wanted so a fresh click
always starts instantly. A recorder that dies after a healthy run (stream
ended at a cam relaunch, crash, failed spawn) is replaced the tick its
death is seen -- measured 0.15 s -- and the relaunch passes --resume:
record from the ring's oldest available frame instead of 'now', so the
ring bridges the handover with no gap. Fresh clicks still record from now.

The GUI has ZERO recording logic left: the checkboxes render the backend's
published flags and send plain toggles. Auto-record also engages the same
tick tracking starts (no GUI round-trip) and works with the GUI closed.

The recorder keeps its simple lifecycle (stream ends -> finalize, exit;
continuity is the backend's job) and now rolls a new file on ANY format
change (shape, mosaic, or bit depth -- not shape alone).

Also: sim cams with --sim-downscale > 1 crashed outright (torch call on
the numpy render output). skysim.render now returns the torch container
tensor; the cam upscales by INTEGER pixel replication (no interpolation,
ADU-exact, edge-replicated remainders) and converts to numpy only at the
shm write egress.

Verified end-to-end headless (full-size sim frames): record grows a .ser,
recording rides through camera disconnect/reconnect, a murdered recorder
is respawned in 0.15 s with --resume and keeps writing, rapid off->on
toggling holds. 9/9.

+++ astrolock/seeker/tests/test_skysim.py
@@ -30,7 +30,9 @@ def test_source_projects_to_boresight():
     # a source exactly at the boresight -> image centre
     dirs = b.view(1, 1, 3).contiguous()
     frame = sim.render(enc_az, enc_alt, 0.0, 0.0, dirs, torch.tensor([2.0]), exposure_s=0.2, substeps=1)
-    win = frame[cy - 8:cy + 9, cx - 8:cx + 9]
+    # render returns the torch container dtype (uint16), for which torch implements few
+    # reductions -- widen for the max/argmax analysis here.
+    win = frame[cy - 8:cy + 9, cx - 8:cx + 9].to(torch.int32)
     assert int(win.max()) > 1000, f"no bright peak at boresight: {int(win.max())}"
     ly, lx = np.unravel_index(int(win.argmax()), win.shape)
     assert abs(lx - 8) <= 2 and abs(ly - 8) <= 2, f"boresight peak off-center at ({lx},{ly})"
@@ -39,7 +41,7 @@ def test_source_projects_to_boresight():
     off = math.radians(0.5)
     d2 = (b * math.cos(off) + A * math.sin(off)).view(1, 1, 3).contiguous()
     frame2 = sim.render(enc_az, enc_alt, 0.0, 0.0, d2, torch.tensor([2.0]), exposure_s=0.2, substeps=1)
-    py, px = np.unravel_index(int(frame2.argmax()), frame2.shape)
+    py, px = np.unravel_index(int(frame2.to(torch.int32).argmax()), frame2.shape)
     exp_x = cx + sim.f_px * math.tan(off)
     assert abs(px - exp_x) <= 3, f"off-axis peak at x={px}, expected {exp_x:.1f}"
     assert abs(py - cy) <= 3, f"off-axis peak drifted in y: {py}"
